+++ modules/facebook_ip_correlation/facebook_ip_correlation/__init__.py
@@ -1,0 +1,6 @@
+"""Facebook IP correlation module."""
+
+from .extractor import CorrelationRecord, IpObservation, extract_observations, process_export
+
+__all__ = ["CorrelationRecord", "IpObservation", "extract_observations", "process_export"]
+